requirements.txt
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
flasgger==0.9.7.1
Flask==3.0.0
Flask-Testing==0.8.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
pymongo==4.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
urllib3==2.0.6
Werkzeug==3.0.0
